dbus-asio: Fix the problems with dbus-asio.pc.in
CMake recipe was good enough to compile/run its own examples, but when I
tried to use it with pkgconfig with libvid/Chromium, it turned out that
pkgconfig file is not good enough. It was failing to find some boost
symbols because boost libraries was not listed in link flags.

Change-Id: Ia6fa223037e72650e0f7810a08c5bcdceea69056