dbus-asio: DBus::Transport constructor should wait until its thread is initialized.
DBus::Native creates a DBus::Transport and DBus::AuthenticationProtocol
instance, and passes Transport to AuthenticationProtocol. This happens
on main thread.

If DBusNative::BeginAuth is called early enough on main thread, it tries
to send the message before m_socket is initialized and first "\" message
is sent.

async_write fails with error, and we silently swallow that.

When we try to send the first message without authentication, dbus
connection gets closed.

Change-Id: I94a57fd4d49927a8065f1136856671fbbb5b4c4e