dbus-asio: Added basic test support using ctest
New tests should be added to tests directory and CMakeLists.txt should
be modified to include the test in build.

To run the test, create a build directory and type:
cmake ../ inside the created directory
make all
make tests

will run the tests. It is not integrated to bitbake(yet).

Change-Id: I79469c4f0bf826c716059ddc3d64a0df1c504a03