dbus-asio: Fix threading and lifetime issues BS#31371
dbus-asio was failing to start up correctly.  Various potential causes
have been found, and the following changes made:

m_Thread startup could notify before wait_for() was ready.  Adjusted order
and added locking. We should not print a diagnostic for a socket error
after we have sent a shutdown.  We should also ensure that we have sent
the shutdown before stopping m_Thread.  Access to m_ReadyToSend should be
protected by the relevant mutex in all cases, otherwise there is a race.

Buffers passed to boost::asio must remain valid until the IO is complete,
so we copy the data to avoid involving the caller in buffer lifetime
management.

Completion handlers are despatched by calling run() or similar on the
io_service_context, which we were not doing.  For now, add yet another
thread to handle this.  We should consider re-writing the transport
thread so that it can handle this as well as the read interface, but this
appears sufficient to make the design function.

Change-Id: I1e86a5e4b4ea7ffd2d06163bbc702b4ff15a2394