dbus_transport: Eliminate "spare" thread
Since we already have an io-worker thread sitting ready to perform
asynchronous operations, it seems a bit silly for us to spin up our
own transport thread and have it make synchronous calls to the asio
API.  Worse, there is some reason to suspect that this may be
causing the asio scheduler state to get messed up.  Eliminate the
extra TransportThread in favour of using async_read() calls and
letting the io-worker dispatch them.

Change-Id: I4d612de9fca747902d9bcdb2078de8ba56dbb55b