dbus-asio: Support abstract sockets
dbus daemons can use abstract sockets. The prefix for the dbus
address for such sockets is "unix:abstract=". The address may
also be appended with a guid. The socket path is obtained by
removing the prefix and the guid parts. A '/0' character is
prepended to the path to indicate that it is an abstract
socket.