dbus-asio: Unit tests for all dbus types
Previous changes ensured that data is always written
in the host byte order. During unmarshalling bytes
are swapped when the host order MATCHES the dbus data
stream order due to how bytes are read in. For double
values this is incorrect as bytes need swapping in all
cases. This has been addressed in this commit. The
dbus_asio_test has also been removed when running
'make test' as it requires a dbus session ID to be
exported.

Change-Id: I818623f5f7f0b1f7ea1f3e54a821ce0ac6da061c