dbus-asio: Only byte swap when host and dbus byte orders mismatch
Dbus types were read in by reading a byte, shifting the previous
value left 8 bits and adding in the new byte. This requires the
bytes to be swapped even if the dbus and host byte order matches.
This commit reads each byte into the correct location in the
dbus values and only swaps the bytes if the byte order does not match.

Change-Id: I5b0490647590a336e22ada9138b8b2eba2a6e48e