dbus-asio: Process incoming data faster
Previously each incoming byte is processed individually
leading to many state machines in Dbus Types and message
protocols. The size of the DBus headers is never very
large and the Dbus message bodies are aready fully cached
in memory before processing. This commit ensures all data
required for unmarshalling is read in and cached before
unmarshalling. This significantly reduces the time to
process Dbus messages.

I subsequently fixed DBus::Type::DictEntry::unmarshall to correctly
parse the value type from the signature. The test case was inadequate to
reveal this problem since it used the same base type for the key and the
value. My test case for this was too invasive to backport to fix up with
this commit, so it will arrive later along with other improved tests.

Change-Id: Ic11bd75954878f723d3222140b66b1adf57842e2