tests: Add DictEntrys test where key and value types are different
An early incarnation of Ic11bd75954878f723d3222140b66b1adf57842e2 failed
to parse the types of a DICT_ENTRY correctly. That was fixed in a later
version, but so much else had changed in subsequent changes that it was
hard to land the changes required for the unit test to reproduce the
bug. Those changes are here.

The existing test is modified to use different base types for key and
value and a new tests are added to use a container type.

Change-Id: Idc91859da66c4dd9688c70119e33fd77c3ff52bd