MessageIStream: Read endianness-swapped double correctly
The implementation of MessageIStream::read needs to honour the
m_swapByteOrder member too. The easiest way to do this is to read a
correctly-byte-swapped uint64_t and then convert it to a double.

The tests in dbus_messageistream_unittest and dbus_type_double_unittest
were clearly wrong but in exactly the same way as the implementation so
this wasn't caught automatically. :(

(I checked that on a PowerPC big-endian machine the bytes in a double
are reversed compared to my x86-64 machine.)

Change-Id: I98fca4f11eac8ee092782a17503f8261b7b79628