Variant: Differentiate type and contained signature
If we use the Base::m_Signature member to contain the type contained in
the variant then we'll be lying to anyone who calls getSignature() to
retrieve the actual type of the variant. Let's make Variant act like all
the other Type classes and report its own type via getSignature and
maintain its contained type separately.

Change-Id: Ic7873d4773ffda4a8260b3acb3c4f19891add863