Message: Marshall signatures correctly
According to the D-Bus specification, a message header consists of:

 BYTE, BYTE, BYTE, BYTE, UINT32, UINT32, ARRAY of STRUCT of (BYTE,VARIANT)

so I was somewhat confused by all the message marshalling code adding
the signature for each parameter directly without marking it as a
variant. This fell foul of some extra checks I added to
DBus::Type::Array to ensure that the members of the array were
homogeneous.

This turns out to be because the implementation of
DBus::Type::Signature::marshall incorrectly adds the type code below a
comment that hints that something fishy is going on:

 // NOTE: We marshall out the 'g' to indicate a signature, but don't parse it on unmarshall

The signature of the signature is being incorrectly emitted as part of
the signature itself!

If the signature is added to the message header correctly as a variant
then the variant will emit the signature of the signature (i.e. the 'g')
and the fishy code is no longer required. This means that
DBus::Type::Signature::marshall becomes the inverse of
DBus::Type::Signature::unmarshall again.

To make this work we need to add another constructor to
DBus::Type::Variant to make it possible to create a variant of a
signature.

Change-Id: I34fd5314b22626dfe3e5aafb8292ab56f9509f92