Fix source code formatting in line with README.first
The README.first file says that the code should be formatted by
clang-format. It looks like this hasn't been done for a while, and all
the recent changes don't match the style. Let's fix that.

Let's also add a script that makes it easier to fix the formatting and
update README.first to point to it.

Change-Id: Idc32fa11d53f35c1d11d0671e6ba7cfd6ca78f64