MessageIStream: Avoid type punning warnings
When GCC is run with -fstrict-aliasing it complains about type punning
from a double to a uint64_t. Let's avoid the whole problem by reading
bytes and using std::reverse to reverse them. This also happens to stop
the function assuming that uint64_t and double are the same size (which
they always are.)