tests: Avoid type punning doubles to uint64_t
When GCC is run with -fstrict-aliasing it complains about type punning
from a double to a uint64_t. The recommended way to avoid this problem
is to use memcpy. The compiler ought to be clever enough to realise that
it doesn't actually need to do any copying.

(This is a deliberately different choice to the implementation in
MessageIStream::read in the hope that it means that we're testing the
implementation correctly.)