dbus_native: Drop pointer to Transport before AuthenticationProtocol
We hold a shared_ptr to Transport and so does AuthenticationProtocol.
By dropping ours first, we ensure that Transport is gone by the time
~AuthenticationProtocol() returns, and there is no risk of Transport
attempting to dispatch any calls to the dead AuthenticationProtocol
object.

This leads to the slightly unusual situation where we want
AuthenticationProtocol to be both constructed and destructed
before Transport, rather than the usual last-in-first-out
model that C++ would give us automatically.  So we must do
the construction explicitly to get the ordering we need.

Change-Id: Ifb8fdfa2bd6e5270d4ad8bdba111095c2e87e02d