dbus-auth: Ensure correct destruction order of data members
When destroying the AuthenticationProtocol we need to ensure that
the Transport thread is stopped before any of the objects that it
might touch are destroyed.  This is, to use the vernacular, all a
bit gnarly because there are multiple classes and shared_ptrs
involved.  The DBusTransport object, which holds the Transport
thread, is retained by two shared_ptrs, one in this class and
one in DBusNative.  Fortunately, AuthenticationProtocol itself
is only ever owned by a unique_ptr inside DBusNative, and we can
rely on the destruction order within DBusNative to guarantee
that DBusNative's shared_ptr<DBusTransport> will be gone before
AuthenticationProtocol starts to be destructed, and hence the
DBusTransport will be destroyed synchronously as soon as we
drop our own shared_ptr to it.

So, if we make sure that this shared_ptr is the last member
of our class, we can be certain that none of the other data
members will be accessed by the transport thread after they've
been destroyed.