Log: Move method bodies out of line
Defining all our methods in the header file means that the static
data member "m_Level" also has to be globally visible.  However it's
subject to interposition which means that it gets a COPY reloc,
and this in turn makes it impossible to build the library with
-fvisibility=protected because the linker won't generate COPY relocs
against protected symbols.

Avoid this whole mess by hiding the definition of all the accessor
methods away in the .cpp file; this in turns means we can make the
global level variable be local to that compilation unit.